requirements.txt
tweepy==4.14.0
google-generativeai==0.8.3
supabase==2.11.0
python-dotenv==1.0.1